requirements.txt
asgiref==3.3.4
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.4
django-admin-sortable2==1.0
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-modeltranslation==0.17.2
django-resized==0.3.11
django-unused-media==0.2.2
idna==3.3
Pillow==8.2.0
psycopg2-binary==2.9.3
pytils==0.3
pytz==2021.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.9
